requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<0.28
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
model-bakery>=1.17,<2.0
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0